AAAAAA

+++ server/include/Server.hpp
@@ -8,7 +8,7 @@ namespace tcp {
 
 using socket_t = int;
 using data_t = std::vector<uint8_t>;
-using error_t = int;
+using error_t = long;
 
 struct ServerConfig {
     using duration = std::chrono::duration<std::uint64_t>;
